Add side menu table view controller and localization files
