+++ Mr-Ride-iOS/Mr-Ride-iOS.xcodeproj/xcuserdata/SK.xcuserdatad/xcschemes/Mr-Ride-iOS.xcscheme
@@ -50,7 +50,8 @@
       ignoresPersistentStateOnLaunch = "NO"
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
-      allowLocationSimulation = "YES">
+      allowLocationSimulation = "YES"
+      language = "en">
       <BuildableProductRunnable
          runnableDebuggingMode = "0">
          <BuildableReference
