+++ Mr-Ride-iOS/Mr-Ride-iOS/TableViewCells/LeftSideMenuTableViewCell.swift
@@ -29,6 +29,7 @@ class LeftSideMenuTableViewCell: UITableViewCell {
         super.setSelected(selected, animated: animated)
 
         // Configure the view for the selected state
+//        leftSideMenuItemLabel.textColor = UIColor.mrWhiteColor()
     }
     
 }
