--- Mr-Ride-iOS/Mr-Ride-iOS/TableViewCells/MyTableViewCell.swift
@@ -1,24 +1,0 @@
-//
-//  MyTableViewCell.swift
-//  Mr-Ride-iOS
-//
-//  Created by SK on 5/23/16.
-//  Copyright © 2016 AppWorks School Snakeking. All rights reserved.
-//
-
-import UIKit
-
-class MyTableViewCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
