--- Mr-Ride-iOS/Mr-Ride-iOS/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  Mr-Ride-iOS
-//
-//  Created by SK on 5/23/16.
-//  Copyright © 2016 AppWorks School Snakeking. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
