+++ Mr-Ride-iOS/Mr-Ride-iOS/zh-Hant.lproj/LaunchScreen.strings
@@ -1,0 +1,1 @@
+
