+++ Mr-Ride-iOS/Mr-Ride-iOS/zh-Hant.lproj/Main.strings
@@ -1,0 +1,6 @@
+
+/* Class = "UIBarButtonItem"; title = "Item"; ObjectID = "RcN-vc-jNR"; */
+"RcN-vc-jNR.title" = "Item";
+
+/* Class = "UIButton"; normalTitle = "Button"; ObjectID = "SvV-ca-wiR"; */
+"SvV-ca-wiR.normalTitle" = "Button";
