+++ Mr-Ride-iOS/zh-Hant.lproj/Localizable.strings
@@ -1,0 +1,8 @@
+/* 
+  Localizable.strings
+  Mr-Ride-iOS
+
+  Created by SK on 5/24/16.
+  Copyright © 2016 AppWorks School Snakeking. All rights reserved.
+*/
+"Let's Ride" = "來騎吧";
